# doc/api.yaml
openapi: 3.0.3
info:
  title: Wasa Photo
  description: This OpenAPI document describes the simplified login API.
  version: '1'
  termsOfService: http://swagger.io/terms/
  contact:
    email: bborsa15@outlook.it
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: https://wasainsta.swagger.io/api/v3
paths:
  /login:
    post:
      tags:
        - Login
      summary: Logs in the user
      description:
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Login"
        required: true
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  identifier:
                    type: integer
                    example: 012345
      security:
         - bearerAuth: []          
...
  /users/{username}/follow/{followid}:
    parameters:
       - $ref: "#/components/parameters/username"
       - $ref: "#/components/parameters/followid"
    put:
        tags: 
          - User
        summary: Follow another user 
        description: this method allows to follow a user of your choice
        operationId: followUser
        requestBody:
          description: Add a new follow
          content:
           application/json:
             schema:
              $ref: '#/components/schemas/User'
        responses:
         '200':
          description: successful operation
          content:
            application/json:
              schema:
               $ref: '#/components/schemas/User'   
         '404':
          description: User not follow
        security:
         - bearerAuth: []      
    delete:
        tags:
         - User
        summary: Delete follow
        description: this method allows to delete a follows to the database
        operationId: unfollowUser
        responses:
          '400':
           description: Invalid ID supplied
          '404':
           description: User not unfollow
        security:
         - bearerAuth: []         
...
  /users/{username}/userprofile:
   parameters:
     - $ref: "#/components/parameters/username"
   get:
         tags:
           - User
         summary: Get user profile by Username
         description: This method allows to obtain the user profile 
         operationId: getUserProfile
         responses:
           '200': 
             description: successful operation
             content:
              application/json:
                schema:
                  $ref: '#/components/schemas/User'          
              application/xml:
                schema:
                  $ref: '#/components/schemas/User'
           '400':
             description: Invalid Username supplied
           '404':
             description: User not found
         security:
          - bearerAuth: []      
...
  /users/{username}/photo/{phid}/like/{likeid}:    
   parameters:
        - $ref: "#/components/parameters/username"
        - $ref : "#/components/parameters/phid"
        - $ref : "#/components/parameters/likeid"
   put:
      tags:
        - Photo
      summary: Put a like 
      description: You can like a photo
      operationId: likePhoto
      requestBody:
        description: That photo was liked
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Photo'
      responses:
        default:
          description: successful operation
      security:
         - bearerAuth: []    
   delete:
      tags:
        - Photo
      summary: Unlike a photo
      description: You can unlike a photo
      operationId: unlikePhoto
     
      responses:
        '400':
          description: Invalid photo supplied
        '404':
          description: Photo not found
      security:
         - bearerAuth: []     
...
  /user/{username}/streamuser:
   parameters: 
       - $ref: "#/components/parameters/username"
   get:
      tags:
        - Photo
      summary: Returns stream by username
      description: Returns a map of photos
      operationId: getMyStream
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: integer
                  format: int32
      security:
         - bearerAuth: []    
...
components:
   securitySchemes:
     bearerAuth:             
       type: http
       scheme: bearer
       bearerFormat: JWT
   parameters:
    username:
      schema:
          description: username
          type: string
          pattern: '^.*?$'
          example: bea
          minLength: 5
          maxLength: 30
      name: username
      in: path
      required: true
      description: "username of user."
    phid:
      schema:
        description: photo id .
        type: string
        pattern: '^.*?$'
        example: "123"
        minLength: 3
        maxLength: 16
      name: phid 
      in: path
      required: true
      description: "photo of user"
...
    likeid:
      schema:
        description: likeiD 
        type: string
        pattern: '^.*?$'
        example: "#L1234"
        minLength: 3
        maxLength: 16
      name: likeid
      in: path
      required: true
      description: "like of photo"
    followid:
      schema:
        description: followid
        type: string
        pattern: '^.*?$'
        example: "#F1234"
        minLength: 3
        maxLength: 16
      name: followid
      in: path
      required: true
      description: "follow of user"
...
   schemas:
      User:
       type: object
       properties:
        identifier:
          description: Unique identifier.
          type: string
          pattern: '^.*?$'
          example: "3737"
          minLength: 3
          maxLength: 16
        username:
          type: string
          example: Bea
      Login:        
        title: Login
        description: Login.
        type: object
        properties:
          username:
           description: user
           type: string
           example: Beatrice
           pattern: '^.*?$'
           minLength: 3
           maxLength: 16
...
      Photo:
       title: Photo
       description: photo
       properties:
        phid: 
          description: photo identifier.
          type: string
          pattern: '^.*?$'
          example: "3445"
          minLength: 3
          maxLength: 16
        userid:
          description: user identifier.
          type: string
          pattern: '^.*?$'
          example: "87866" 
          minLength: 3
          maxLength: 16
        fileid: 
          description: file
          type: string
          format: binary
          minLength: 3
          maxLength: 16
        dateid:
          description: date.
          type: string
          format: date-time
          example: "2022-11-02T09:00:00Z"
          minLength: 3
          maxLength: 9999
        likeInc:
          description: likes count
          type: integer
          example: 0
          minimum: 0
          maximum: 9999
        commentInc:
          description: photo count
          type: integer
          example: 0
          minimum: 0
          maximum: 9999
        likestate:
          description: like status
          type: boolean
          example: true